Create archive sub-directories when required
When generating a document containing embedded images, an IOError
would be raised when copying the image file into the Pictures/
sub-directory of the archive because the directory would not exist.

This commits allows the backend to verify the directory's existence on
creation, and creates it if needed. This fixes dagwieers/asciidoc-odf#42.